backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Google AI (Gemini)
google-generativeai==0.8.3

# Qdrant Vector Database
qdrant-client==1.12.1

# Authentication & Security
pyjwt==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment Variables
python-dotenv==1.0.1

# HTTP Requests
httpx==0.27.2
requests==2.32.3

# Data Validation
pydantic==2.10.2
pydantic-settings==2.6.1

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.9.0
